Bump Kubernetes deps, buf, and provider-aws
These all had quite a lot of interactions so I had to do them together.

Note we can't bump to Kubernetes v0.31.0 because it's newer than what
the latest crossplane-runtime release is using. We'll need to update
crossplane-runtime first.

I'm also seriously questioning keeping the dependency on provider-aws.
It's massive and has a lot of painful dependencies to fix when we update
the import. On the other hand, I experimented with e.g. using
provider-nop or provider-kubernetes for examples and didn't like it.

Signed-off-by: Nic Cope <nicc@rk0n.org>